requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
SQLAlchemy==2.0.28 
Werkzeug==3.0.1
python-dotenv==1.0.0
xhtml2pdf==0.2.13
reportlab==4.0.7
html5lib==1.1
Pillow==10.4.0
python-socketio==5.11.3
python-engineio==4.8.0
gunicorn==21.2.0
eventlet==0.33.3
psycopg2-binary==2.9.9
